requirements.txt
python-telegram-bot
pydantic
pydantic_settings
sqlalchemy
pixivpy3
gallery-dl